@@ -34,12 +34,6 @@ const _horizontalDesktopPadding = 81.0;
 const _carouselHeightMin = 200.0 + 2 * _carouselItemMargin;
 const _desktopCardsPerPage = 4;
 
-const _shrineTitle = 'Shrine';
-const _rallyTitle = 'Rally';
-const _craneTitle = 'Crane';
-const _homeCategoryMaterial = 'MATERIAL';
-const _homeCategoryCupertino = 'CUPERTINO';
-
 class ToggleSplashNotification extends Notification {}
 
 class HomePage extends StatelessWidget {
@@ -47,10 +41,11 @@ class HomePage extends StatelessWidget {
   Widget build(BuildContext context) {
     var carouselHeight = _carouselHeight(.7, context);
     final isDesktop = isDisplayDesktop(context);
+    final localizations = GalleryLocalizations.of(context);
+    final studyDemos = studies(localizations);
     final carouselCards = <Widget>[
       _CarouselCard(
-        title: _shrineTitle,
-        subtitle: GalleryLocalizations.of(context).shrineDescription,
+        demo: studyDemos['shrine'],
         asset: AssetImage('assets/studies/shrine_card.png'),
         assetColor: const Color(0xFFFEDBD0),
         assetDark: AssetImage('assets/studies/shrine_card_dark.png'),
@@ -59,8 +54,7 @@ class HomePage extends StatelessWidget {
         studyRoute: ShrineApp.loginRoute,
       ),
       _CarouselCard(
-        title: _rallyTitle,
-        subtitle: GalleryLocalizations.of(context).rallyDescription,
+        demo: studyDemos['rally'],
         textColor: RallyColors.accountColors[0],
         asset: AssetImage('assets/studies/rally_card.png'),
         assetColor: const Color(0xFFD1F2E6),
@@ -69,8 +63,7 @@ class HomePage extends StatelessWidget {
         studyRoute: RallyApp.loginRoute,
       ),
       _CarouselCard(
-        title: _craneTitle,
-        subtitle: GalleryLocalizations.of(context).craneDescription,
+        demo: studyDemos['crane'],
         asset: AssetImage('assets/studies/crane_card.png'),
         assetColor: const Color(0xFFFBF6F8),
         assetDark: AssetImage('assets/studies/crane_card_dark.png'),
@@ -79,8 +72,7 @@ class HomePage extends StatelessWidget {
         studyRoute: CraneApp.defaultRoute,
       ),
       _CarouselCard(
-        title: fortnightlyTitle,
-        subtitle: GalleryLocalizations.of(context).fortnightlyDescription,
+        demo: studyDemos['fortnightly'],
         asset: AssetImage('assets/studies/fortnightly_card.png'),
         assetColor: Colors.white,
         assetDark: AssetImage('assets/studies/fortnightly_card_dark.png'),
@@ -88,8 +80,7 @@ class HomePage extends StatelessWidget {
         studyRoute: FortnightlyApp.defaultRoute,
       ),
       _CarouselCard(
-        title: GalleryLocalizations.of(context).starterAppTitle,
-        subtitle: GalleryLocalizations.of(context).starterAppDescription,
+        demo: studyDemos['starterApp'],
         asset: AssetImage('assets/studies/starter_card.png'),
         assetColor: const Color(0xFFFAF6FE),
         assetDark: AssetImage('assets/studies/starter_card_dark.png'),
@@ -102,24 +93,26 @@ class HomePage extends StatelessWidget {
     if (isDesktop) {
       final desktopCategoryItems = <_DesktopCategoryItem>[
         _DesktopCategoryItem(
-          title: _homeCategoryMaterial,
+          category: GalleryDemoCategory.material,
           asset: AssetImage('assets/icons/material/material.png'),
-          demos: materialDemos(context),
+          demos: materialDemos(localizations),
         ),
         _DesktopCategoryItem(
-          title: _homeCategoryCupertino,
+          category: GalleryDemoCategory.cupertino,
           asset: AssetImage('assets/icons/cupertino/cupertino.png'),
-          demos: cupertinoDemos(context),
+          demos: cupertinoDemos(localizations),
         ),
         _DesktopCategoryItem(
-          title: GalleryLocalizations.of(context).homeCategoryReference,
+          category: GalleryDemoCategory.other,
           asset: AssetImage('assets/icons/reference/reference.png'),
-          demos: referenceDemos(context),
+          demos: otherDemos(localizations),
         ),
       ];
 
       return Scaffold(
         body: ListView(
+          // Makes integration tests possible.
+          key: ValueKey('HomeListView'),
           padding: EdgeInsetsDirectional.only(
             top: isDesktop ? firstHeaderDesktopTopPadding : 21,
           ),
@@ -323,9 +316,12 @@ class _AnimatedHomePageState extends State<_AnimatedHomePage>
 
   @override
   Widget build(BuildContext context) {
+    final localizations = GalleryLocalizations.of(context);
     return Stack(
       children: [
         ListView(
+          // Makes integration tests possible.
+          key: ValueKey('HomeListView'),
           children: [
             SizedBox(height: 8),
             Container(
@@ -344,27 +340,36 @@ class _AnimatedHomePageState extends State<_AnimatedHomePage>
               startDelayFraction: 0.00,
               controller: _animationController,
               child: CategoryListItem(
-                title: _homeCategoryMaterial,
+                key: PageStorageKey<GalleryDemoCategory>(
+                  GalleryDemoCategory.material,
+                ),
+                category: GalleryDemoCategory.material,
                 imageString: 'assets/icons/material/material.png',
-                demos: materialDemos(context),
+                demos: materialDemos(localizations),
               ),
             ),
             _AnimatedCategoryItem(
               startDelayFraction: 0.05,
               controller: _animationController,
               child: CategoryListItem(
-                title: _homeCategoryCupertino,
+                key: PageStorageKey<GalleryDemoCategory>(
+                  GalleryDemoCategory.cupertino,
+                ),
+                category: GalleryDemoCategory.cupertino,
                 imageString: 'assets/icons/cupertino/cupertino.png',
-                demos: cupertinoDemos(context),
+                demos: cupertinoDemos(localizations),
               ),
             ),
             _AnimatedCategoryItem(
               startDelayFraction: 0.10,
               controller: _animationController,
               child: CategoryListItem(
-                title: GalleryLocalizations.of(context).homeCategoryReference,
+                key: PageStorageKey<GalleryDemoCategory>(
+                  GalleryDemoCategory.other,
+                ),
+                category: GalleryDemoCategory.other,
                 imageString: 'assets/icons/reference/reference.png',
-                demos: referenceDemos(context),
+                demos: otherDemos(localizations),
               ),
             ),
           ],
@@ -393,12 +398,12 @@ class _AnimatedHomePageState extends State<_AnimatedHomePage>
 
 class _DesktopCategoryItem extends StatelessWidget {
   const _DesktopCategoryItem({
-    this.title,
+    this.category,
     this.asset,
     this.demos,
   });
 
-  final String title;
+  final GalleryDemoCategory category;
   final ImageProvider asset;
   final List<GalleryDemo> demos;
 
@@ -416,7 +421,7 @@ class _DesktopCategoryItem extends StatelessWidget {
           child: Column(
             children: [
               _DesktopCategoryHeader(
-                title: title,
+                category: category,
                 asset: asset,
               ),
               Divider(
@@ -430,6 +435,8 @@ class _DesktopCategoryItem extends StatelessWidget {
                   removeTop: true,
                   context: context,
                   child: ListView(
+                    // Makes integration tests possible.
+                    key: ValueKey('${category.name}DemoList'),
                     children: [
                       const SizedBox(height: 12),
                       for (GalleryDemo demo in demos)
@@ -451,16 +458,18 @@ class _DesktopCategoryItem extends StatelessWidget {
 
 class _DesktopCategoryHeader extends StatelessWidget {
   const _DesktopCategoryHeader({
-    this.title,
+    this.category,
     this.asset,
   });
-  final String title;
+  final GalleryDemoCategory category;
   final ImageProvider asset;
 
   @override
   Widget build(BuildContext context) {
     ColorScheme colorScheme = Theme.of(context).colorScheme;
     return Material(
+      // Makes integration tests possible.
+      key: ValueKey('${category.name}CategoryHeader'),
       color: colorScheme.onBackground,
       child: Row(
         children: [
@@ -483,7 +492,7 @@ class _DesktopCategoryHeader extends StatelessWidget {
               child: Semantics(
                 header: true,
                 child: Text(
-                  title,
+                  category.displayTitle(GalleryLocalizations.of(context)),
                   style: Theme.of(context).textTheme.headline5.apply(
                         color: colorScheme.onSurface,
                       ),
@@ -933,8 +942,7 @@ class _DesktopPageButton extends StatelessWidget {
 class _CarouselCard extends StatelessWidget {
   const _CarouselCard({
     Key key,
-    this.title,
-    this.subtitle,
+    this.demo,
     this.asset,
     this.assetDark,
     this.assetColor,
@@ -943,8 +951,7 @@ class _CarouselCard extends StatelessWidget {
     this.studyRoute,
   }) : super(key: key);
 
-  final String title;
-  final String subtitle;
+  final GalleryDemo demo;
   final ImageProvider asset;
   final ImageProvider assetDark;
   final Color assetColor;
@@ -961,6 +968,8 @@ class _CarouselCard extends StatelessWidget {
     final textColor = isDark ? Colors.white.withOpacity(0.87) : this.textColor;
 
     return Container(
+      // Makes integration tests possible.
+      key: ValueKey(demo.describe),
       margin:
           EdgeInsets.all(isDisplayDesktop(context) ? 0 : _carouselItemMargin),
       child: Material(
@@ -993,13 +1002,13 @@ class _CarouselCard extends StatelessWidget {
                   mainAxisAlignment: MainAxisAlignment.end,
                   children: [
                     Text(
-                      title,
+                      demo.title,
                       style: textTheme.caption.apply(color: textColor),
                       maxLines: 3,
                       overflow: TextOverflow.visible,
                     ),
                     Text(
-                      subtitle,
+                      demo.subtitle,
                       style: textTheme.overline.apply(color: textColor),
                       maxLines: 5,
                       overflow: TextOverflow.visible,
@@ -1057,6 +1066,7 @@ class _StudyWrapperState extends State<StudyWrapper> {
                 button: true,
                 excludeSemantics: true,
                 child: FloatingActionButton.extended(
+                  key: ValueKey('Back'),
                   onPressed: () {
                     Navigator.of(context)
                         .popUntil((route) => route.settings.name == '/');
